LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/textfieldselection/selection-not-application-textarea-expected.txt


FAIL text field selection for the input textarea assert_equals: updated selectionStart expected 1 but got 0
